Add autoscaler cron to async api
